requirements.txt
streamlit==1.27.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
pydantic==2.4.2
tqdm==4.66.1
plotly==5.18.0
pytest==7.4.3
